requirements.txt
discord.py
python-dotenv
fastapi
uvicorn
aiohttp